docs(course): Update project initialization section in zig command documentation
Updated the documentation for the zig init command, added newly created files build.zig.zon and root.zig , and improved the directory structure comments. Also corrected the file list description to match the actual generated files.Actually,I found that even with zig 0.14.0 (the previous version which also mentioned on this site page),the output of zig init is the same as with zig 0.14.1,which is different from what this documentation descirbes.Anyway,I think it needs to be updated to avoid being misleading.